requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ollama==0.4.4
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-telegram-bot==21.9
sniffio==1.3.1
typing_extensions==4.12.2
